+++ packages/mask/src/plugins/Avatar/Services/bind.ts
@@ -1,20 +1,20 @@
 import { ChainId, isSameAddress } from '@masknet/web3-shared-evm'
-import { NFT_AVATAR_GUN_SERVER } from '../constants'
 import { NetworkPluginID } from '@masknet/plugin-infra'
 import addSeconds from 'date-fns/addSeconds'
 import isBefore from 'date-fns/isBefore'
 import { KeyValue } from '@masknet/web3-providers'
+import { NFT_AVATAR_DB_NAME } from '../constants'
 
-const NFTAvatarDB = KeyValue.createJSON_Storage(NFT_AVATAR_GUN_SERVER)
+const NFTAvatarDB = KeyValue.createJSON_Storage(NFT_AVATAR_DB_NAME)
 
 const cache = new Map<string, [Promise<string | undefined>, Date]>()
 
-async function _getUserAddress(userId: string, networkPluginId?: NetworkPluginID, chainId?: ChainId) {
+async function _getUserAddress(userId: string, networkPluginId?: NetworkPluginID, chainId?: number) {
     const result = await NFTAvatarDB.get<Record<string, string>>(userId)
     return result?.[`${networkPluginId ?? NetworkPluginID.PLUGIN_EVM}-${chainId ?? ChainId.Mainnet}`]
 }
 
-export async function getUserAddress(userId: string, networkPluginId?: NetworkPluginID, chainId?: ChainId) {
+export async function getUserAddress(userId: string, networkPluginId?: NetworkPluginID, chainId?: number) {
     let c = cache.get(userId)
     if (!c || isBefore(new Date(), c[1])) {
         cache.set(userId, [_getUserAddress(userId, networkPluginId, chainId), addSeconds(new Date(), 60)])
@@ -27,7 +27,7 @@ export async function setUserAddress(
     userId: string,
     address: string,
     networkPluginId?: NetworkPluginID,
-    chainId?: ChainId,
+    chainId?: number,
 ) {
     try {
         await NFTAvatarDB.set<Record<string, string>>(userId, {
