+++ packages/mask/src/plugins/Avatar/constants.ts
@@ -1,7 +1,7 @@
 import { PluginId } from '@masknet/plugin-infra'
 
 export const NFT_AVATAR_JSON_SERVER = 'https://configuration.r2d2.to/com.maskbook.avatar.json'
-export const NFT_AVATAR_GUN_SERVER = 'com.maskbook.user'
+export const NFT_AVATAR_DB_NAME = 'com.maskbook.user'
 
 export const NFT_CONTRACT_JSON_VERIFIED_SERVER = 'https://configuration.r2d2.to/com.maskbook.verified_nft.json'
 export const PLUGIN_ID = PluginId.Avatar
